Update vtkbone to build as a VTK 9 module
This is a major change, but is necessary in order to use Python
wrapping with VTK 9.  Since VTK 9 relies on some new CMake tricks,
the vtkbone package export has also changed.  Now there is no
VTKBONE_INCLUDE_DIRS in the export, since the include directories
are now automatically set when linking to vtkbone.